requirements.txt
langchain
langchain_community
langchain_huggingface
pymupdf
tiktoken
rank_bm25
faiss-cpu
langchain-google-genai
google.generativeai
python-dotenv
streamlit
requests
uvicorn
fastapi